requirements.txt
Pillow==8.0.1
requests==2.25.0